unified internal and external plugins dyncfg API
