+++ collectors/systemd-journal.plugin/systemd-journal-dyncfg.c
@@ -2,7 +2,7 @@
 
 #include "systemd-internals.h"
 
-int systemd_journal_directories_dyncfg_cb(const char *transaction, const char *id, DYNCFG_CMDS cmd, BUFFER *payload, usec_t *stop_monotonic_ut, bool *cancelled, void *data) {
+int systemd_journal_directories_dyncfg_cb(const char *transaction, const char *id, DYNCFG_CMDS cmd, BUFFER *payload, usec_t *stop_monotonic_ut, bool *cancelled, BUFFER *result, void *data) {
     CLEAN_BUFFER *action = buffer_create(100, NULL);
     dyncfg_cmds2buffer(cmd, action);
